pyproject.toml
[project]
name = "behavioral-experiments"
version = "0.1.0"
description = "LLM Behavioral Experiments - Game Theory & Decision Making"
authors = [
    {name = "AI Safety Research", email = "research@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pandas>=1.5.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.11.0",
    "numpy>=1.21.0",
    "openai>=1.93.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["runner", "openrouter_client", "database", "analyze", "visualize", "generate_report"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
